+++ v2/core/config.lua
@@ -9,11 +9,12 @@ local LocalPlayer = Services.LocalPlayer
 
 local ConfigModule = {}
 
-ConfigModule.SCRIPT_BUILD_COMMIT = "v2.8.5"
+ConfigModule.SCRIPT_BUILD_COMMIT = "v2.8.6"
 
 -- 1. Full Config Table from backup.lua
 ConfigModule.Config = {
     -- Câu Cá Cốt Lõi
+    AutoMinigame = true,
     AutoCast = false,
     CastDelay = 1.0,
     CastPower = 100,